Add LARColorFrame for mapping; move larDeepCopy into the library
Mapping needs full color (COLMAP reconstruction) while localization needs
only grayscale luma (CV_8UC1), so a single owned-capture type doesn't fit
both. Model them as two composed value types that each carry a LARFrame plus
exactly the pixels their consumer needs:

- LARImageFrame (existing): grayscale luma -> localization.
- LARColorFrame (new): owns a deep-copied color CVPixelBuffer -> mapping.

Both copy the pixels synchronously so the ARFrame returns to ARKit's pool
immediately (avoiding "ARSessionDelegate is retaining N ARFrames"). This makes
snap() symmetric with localize(): capture one owned frame, hand it to a
detached task. LARColorFrame takes the map-local transform explicitly because
that conversion needs the scene graph and differs from localization's raw VIO
pose; the frame stores it as extrinsics.

Add LARMapper.addFrame(_:) to consume a LARColorFrame, and move the
CVPixelBuffer.larDeepCopy utility from the LARScan example into the LocalizeAR
library (kept public with the lar prefix to avoid colliding with other
CVPixelBuffer extensions), so LARColorFrame can use it. The example's copy is
removed to avoid an ambiguous duplicate extension.
